+++ src/interface/Config.ts
@@ -1,0 +1,5 @@
+interface Config {
+    GitHub_Emojis_List: string,
+    Unicode_Emojis_Data: string,
+    Json_File_Name?: string
+}