+++ src/interface/GithubEmoji.ts
@@ -1,0 +1,3 @@
+interface GithubEmoji {
+    [shortcode: string]: string
+}