+++ src/interface/IConfig.ts
@@ -1,4 +1,0 @@
-interface IConfig {
-    GitHub_Emojis_List: string,
-    Unicode_Emojis_Data: string
-}