+++ src/interface/IGithubEmoji.ts
@@ -1,3 +1,0 @@
-interface IGithubEmoji {
-    [shortcode: string]: string
-}